various changes = dj/190 (#281)
ENCLONE VISUAL TESTING
- convert some tests from manual to automated
- image comparison now shows three panels, the third being the diff
- add verbose option
- restore and expand some of the delicate image comparison, because it turns out that we need it
ENCLONE VISUAL OTHER
- print some memory stats on console
- add console button
- update to do list
- work in progress on history management
- don't truncate error messages
- reduce number of displayed clonotype groups from 100 to 50
- tweak help
CRATES
- drop some unneeded features from crates
- enclone no longer depends on bio
- update crates
OTHER
- add check for size file to downloads test
- major rejiggering of argument processing so that PRE is known before SOURCE
STATUS
- ./test, enclone_test, repeat_test_vis and enclone visual manual tests all pass

+++ enclone_main/src/stop.rs
@@ -245,7 +245,7 @@ pub fn main_enclone_stop(mut inter: EncloneIntermediates) -> Result<EncloneState
 
     let mut svgs = Vec::<String>::new();
     let mut group_pics = Vec::<String>::new();
-    let mut last_widths = Vec::<usize>::new();
+    let mut last_widths = Vec::<u32>::new();
     let mut summary = String::new();
     tail_code(
         &tall,
